protocol: make RedirectChannel compatible with PM 28.5
The sketchy meta-redirect previously used was no longer working with
Palemoon 28.5. This is an even more sketchy version that uses a
JavaScript redirect, but seems to work.

This redirect screws up the history, so this introduces a new bug as
well :/

My XPCOM-fu isn't quite good enough to figure out a nice way to
redirect a channel without this crap, so open to suggestions on a more
elegant way to solve this.

BUG=madand#9
TEST=visited dactyl://help-tag/starting and got to the correct page

Signed-off-by: Jack Rosenthal <jack@rosenth.al>